Changed price type in Order from BigDecimal to Long.

+++ API/src/main/java/com/packshop/api/modules/shopping/dto/ProductItemDTO.java
@@ -1,7 +1,5 @@
 package com.packshop.api.modules.shopping.dto;
 
-import java.math.BigDecimal;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,6 +10,6 @@ import lombok.NoArgsConstructor;
 public class ProductItemDTO {
     private Long id;
     private String name;
-    private BigDecimal price;
+    private Long price;
     private String thumbnail;
 }
